+++ packages/act-rules/test/url.spec.ts
@@ -11,7 +11,7 @@ describe('URL evaluation', function () {
   it('Evaluates url', async function () {
     this.timeout(0);
 
-    const url = 'https://www.dgeec.medu.pt/pedidodados';
+    const url = 'https://observatorio.acessibilidade.leadershipbt.com/directories/1';
     const response = await fetch(url);
     const sourceCode = await response.text();