+++ packages/core/src/index.ts
@@ -2,6 +2,5 @@ export * from './qualweb';
 export * from './lib/evaluation';
 export * from './lib/i18n';
 
-// TODO: consider what else should be exposed to other callers.
-export * from './lib/QualwebOptions';
+export { QualwebOptions, PuppeteerLifeCycleEvent } from './lib/QualwebOptions';
 export { CounterReport } from './lib/evaluation/CounterReport';