+++ packages/wcag-techniques/test/techniques/QW-WCAG-T14.spec.ts
@@ -1,0 +1,4 @@
+import tests from '../fixtures/QW-WCAG-T14_test';
+import { buildTest } from './test_template';
+
+buildTest('QW-WCAG-T14', tests);