+++ packages/act-rules/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.6.39
+
+### Patch Changes
+
+- b368b355: Escape selector in R34
+
+## 0.6.38
+
+### Patch Changes
+
+- 91218960: R36 stops throwing when headers attribute is empty
+
 ## 0.6.37
 
 ### Patch Changes