+++ packages/crawler/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.3.20
+
+### Patch Changes
+
+- 2b34d754: Update crawl filter
+
+## 0.3.19
+
+### Patch Changes
+
+- b5b79f1f: Add zip to not crawl list
+
 ## 0.3.18
 
 ### Patch Changes